refactor: Introduce new configuration classes for data processing and file upload, enhancing validation and default settings for improved data handling
