+++ .github/workflows/auto-tag-version.yml
@@ -20,7 +20,7 @@ jobs:
     - name: Get last commit message
       id: last_commit
       run: |
-        message=$(git log -1 --pretty=%B)
+        message=$(git log -1 --pretty=%B | tr '\n' ' ' | sed 's/[\\"]/\\&/g')
         echo "commit_message=${message}" >> $GITHUB_OUTPUT
 
     - name: Get version from csproj
