+++ .github/workflows/version-bump.yml
@@ -16,12 +16,19 @@ jobs:
       pull-requests: write
 
     steps:
-    - name: Checkout code
+    - name: Checkout main branch
       uses: actions/checkout@v4
       with:
         token: ${{ secrets.GITHUB_TOKEN }}
         fetch-depth: 0
         ref: main
+        clean: true
+
+    - name: Ensure up-to-date main
+      run: |
+        git fetch origin main
+        git checkout main
+        git reset --hard origin/main
 
     - name: Setup .NET
       uses: actions/setup-dotnet@v4
@@ -112,6 +119,7 @@ jobs:
 
     - name: Create version bump branch and commit
       if: steps.check-changes.outputs.has-changes == 'true'
+      id: create-version-bump-branch
       run: |
         git config --local user.email "action@github.com"
         git config --local user.name "GitHub Action"
